Using size_t instead of unsigned int for conversions to UTF16 for 64-bit compatibility and to lessen the number of casts.

+++ win32/PlatWin.cxx
@@ -331,7 +331,7 @@ FontCached::FontCached(const FontParameters &fp) :
 		IDWriteTextFormat *pTextFormat;
 		const int faceSize = 200;
 		WCHAR wszFace[faceSize];
-		UTF16FromUTF8(fp.faceName, static_cast<unsigned int>(strlen(fp.faceName))+1, wszFace, faceSize);
+		UTF16FromUTF8(fp.faceName, strlen(fp.faceName)+1, wszFace, faceSize);
 		FLOAT fHeight = fp.size;
 		DWRITE_FONT_STYLE style = fp.italic ? DWRITE_FONT_STYLE_ITALIC : DWRITE_FONT_STYLE_NORMAL;
 		HRESULT hr = pIDWriteFactory->CreateTextFormat(wszFace, NULL,
@@ -486,7 +486,7 @@ public:
 	TextWide(const char *s, int len, bool unicodeMode, int codePage=0) :
 		VarBuffer<wchar_t, stackBufferLength>(len) {
 		if (unicodeMode) {
-			tlen = UTF16FromUTF8(s, len, buffer, len);
+			tlen = static_cast<int>(UTF16FromUTF8(s, len, buffer, len));
 		} else {
 			// Support Asian string display in 9x English
 			tlen = ::MultiByteToWideChar(codePage, 0, s, len, buffer, len);
